Add support for the expert PPDs of Gutenprint
Before, we used only the simplified PPDs of Gutenprint, as the number
of vendor-specific options/attributes in PAPPL was hard-limited to 32,
whereas a Gutenprint PPD for an Epson printer can have ~120 options.

As this limit is simply defined by the constant PAPPL_MAX_VENDOR in
PAPPL's source code, one can easily raise it if one uses one's own
PAPPL build, for example in a Snap.

Therefore we make it now depending on the value of PAPPL_MAX_VENDOR
now whether we use the simplified or the expert PPDs, exactly we use
the expert PPDs if PAPPL_MAX_VENDOR is 256 or higher. snapcraft.yaml
now changes PAPPL_MAX_VENDOR to 256 when building PAPPL.

This also means that from now on the Gutenprint Printer Application
Snap in the Snap Store uses the expert PPD files.