Fix bugs caused by SSL transition in CUPS and PAPPL
Currently, there is work-in-progress on using SSL instead of GNU-TLS
in CUPS and PAPPL. Therefore there are some bugs in the code which we
work around with this commit.

1. On "./configure" for CUPS we use "--with-tls=gnutls" now, as
current CUPS defaults to SSL and this is buggy, causing a segfault
when serving out a HTTPS web interface page.

2. We lock the PAPPL snapshot to use an earlier commit ("Redirect to
HTTPS when the user clicks on Use My Position.", from March 16, 2022,
74962172f0), as the top of the GIT is a work-in-progress/save-work
commit, killing off the TLS-certificate buttons without providing a
(SSL) replacement.