Updated for fixes in PAPPL and removal of libfontembed In PAPPL the bug of text/string-type vendor options
https://github.com/michaelrsweet/pappl/issues/142

got fixed and pappl-retrofit got appropriately updated.

Therefore the patches on PAPPL and pappl-retrofit got removed now.

The installable accessory configuration is now saved in separate
configuration files and not in the state file any more. Therefore we
set the STATE_DIR environment variable in the wrapper script
run-ps-printer-app now.

Also PAPPL's failure of installing subscription.h got fixed now so
that we do not need to install it manually.

In addition libfontembed got merged into libcupsfilters, so we do not
need to build and install it separately any more.