Add versioning for the Snap/Ghostscript 10.01.2 security update
The version number of the Snap is now the upstream version number of
the included Gutenprint plus a package release number which gets
bumped for any change outside Gutenprint.

Current version number is 5.3.4-1

- We allow defining the macros for the version string and its numeric
  components in gutenprint-printer-app.c externally and only fall back
  to version "1.0" if no external definitions have been done.

- In the Makefile we also allow the external definition of a version
  string and numeric components. If only a version string is supplied,
  the numerical components get extracted from the version string.

- In snapcraft.yaml we extract the version number from the upstream
  source code of Gutenprint and assign this number plus a manually set
  package release number to the Snap. We grab this version string of
  the Snap and define it as the version of gutenprint-printer-app in
  the "gutenprint-printer-app" part.

Now we can track versions and promote our Snap to "stable" which is
needed for including it in OS distributions, like Ubuntu 23.10.

We also fix security vulnerability CVE-2023-36664 of Ghostscript by
updating to upstream version 10.01.2..

https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2023-36664